Refactor state machine for Android
