+++ packages/expo-updates/android/src/main/java/expo/modules/updates/UpdatesInterface.kt
@@ -10,7 +10,8 @@ import java.io.File
 
 // this unused import must stay because of versioning
 /* ktlint-disable no-unused-imports */
-import expo.modules.updates.UpdatesConfiguration
+import expo.modules.updates.statemachine.UpdatesStateMachine
+
 /* ktlint-enable no-unused-imports */
 
 /**
