+++ packages/expo-updates/android/src/main/java/expo/modules/updates/UpdatesService.kt
@@ -13,8 +13,8 @@ import java.io.File
 
 // these unused imports must stay because of versioning
 /* ktlint-disable no-unused-imports */
-import expo.modules.updates.UpdatesConfiguration
-import expo.modules.updates.UpdatesController
+import expo.modules.updates.statemachine.UpdatesStateMachine
+
 /* ktlint-enable no-unused-imports */
 
 /**
