+++ packages/expo-updates/android/src/main/java/expo/modules/updates/statemachine/UpdatesStateChangeEventSender.kt
@@ -1,0 +1,10 @@
+package expo.modules.updates.statemachine
+
+/**
+Protocol with a method for sending state change events to JS.
+In production, this is implemented by the UpdatesController singleton.
+ */
+interface UpdatesStateChangeEventSender {
+  // Method to send events
+  fun sendUpdateStateChangeEventToBridge(eventType: UpdatesStateEventType, fields: List<String>, values: Map<String, Any>)
+}
