+++ packages/expo-updates/android/src/main/java/expo/modules/updates/statemachine/UpdatesStateError.kt
@@ -1,0 +1,16 @@
+package expo.modules.updates.statemachine
+
+/**
+ * For the state machine, errors are stored as data objects.
+ * For now, we just have the "message" property.
+ */
+data class UpdatesStateError(
+  var message: String
+){
+  val json: MutableMap<String, String>
+    get() {
+      return mutableMapOf(
+        "message" to message
+      )
+    }
+}
