+++ packages/expo-updates/android/src/main/java/expo/modules/updates/statemachine/UpdatesStateValue.kt
@@ -1,0 +1,11 @@
+package expo.modules.updates.statemachine
+
+/**
+All the possible states the machine can take.
+ */
+enum class UpdatesStateValue(val value: String) {
+  Idle("idle"),
+  Checking("checking"),
+  Downloading("downloading"),
+  Restarting("restarting")
+}
